requirements.txt
python-pptx
groq
flask
lxml
python-docx
openpyxl
